refactor(header, footer): add-info, register, participate은 footer X, header categoryTab X
