+++ src/components/layouts/footer/ConditionalFooter.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { usePathname } from 'next/navigation';
-import Footer from './footer';
+import Footer from './Footer';
 
 export default function ConditionalFooter() {
   const pathname = usePathname();
